Rename & remove data.js
